+++ packages/orq-rc/RELEASES.md
@@ -8899,3 +8899,13 @@ Based on:
 - [python v4.11.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.39 - packages/orq-rc
+
+## 2026-06-22 14:05:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.784.0 (2.911.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.40 - packages/orq-rc